+++ CHANGELOG.md
@@ -243,6 +243,9 @@
     `/FEATURES` argument.
   - Cleaned up a bit in the tests, removed excessive piping.
   - Fixed minor typo in examples.
+  - Changes to integration tests
+    - CONN feature was temporarily removed from the instances installed by the
+      integration tests. This is due to issue #1105.
 - Changes to SqlRS
   - Updated the integration tests to stop the Reporting Services service after
     the integration test. This is to save memory on the AppVeyor build worker.